Split mevent_client from mevent_server
